Separate statement and expressions' helper functions.
