+++ compiler/ast/statements_test_helpers.go
@@ -191,25 +191,3 @@ func (ds *DefStatement) ShouldHasSplatParam(t *testing.T, paramName string) {
 
 	t.Fatalf("Can't find splat param '%s' in method '%s'", paramName, ds.Name.Value)
 }
-
-/*
-AssignExpression
-*/
-
-// NameIs compares the assignment's variable name and expected name
-func (ae *AssignExpression) NameIs(n string) bool {
-	return ae.Variables[0].(testNode).NameIs(n)
-}
-
-/*
-Identifier
-*/
-
-// NameIs compares the identifier's name and expected name
-func (i *Identifier) NameIs(n string) bool {
-	if i.Value == n {
-		return true
-	}
-
-	return false
-}
